Fail early if new query model operations aren't added to the gentests
This fails at import time so the error message is a bit messy, but it
acts as a quick and dirty check when new query model operations are
added. Note that just because there is a stategy for an operation
(which this change ensures) that doesn't mean that the strategy
actually forms part of the effective strategy tree. There is a
separate check that all the operations are actually generated in
practice, but that is slow to run so we can't include it in every
build.