Handle relative module paths under Python 3.9.5
Python 3.9.5 included the [change][1]:

 * bpo-43105: Importlib now resolves relative paths when creating module
   spec objects from file locations.

This was decided to be a mistake and [backed out][2] in Python 3.9.6:

 * bpo-44070: No longer eagerly makes import filenames absolute, except
   for extension modules, which was introduced in 3.9.5.

However the current Data Builder Docker image uses 3.9.5 so we need to
support it. And it's better to be robust against these kinds of change
in behaviour in any case. So instead of converting the module path to a
string ourselves, we ask Python what it thinks the module path is as
that's ultimately the thing we'll be checking against.

[1]: https://docs.python.org/3.9/whatsnew/changelog.html#id54
[2]: https://docs.python.org/3.9/whatsnew/changelog.html#id43