fix: Work around SQLAlchemy bug compiling SQL to strings
A particular combination of SQLAlchemy features (`bind_expressions`,
"expanding" BindParameters, and `literal_binds` compilation) exposes a
bug in SQLAlchemy.

SQLAlchemy are quite clear that compiling queries to strings with
literal values in (rather than parameter placeholders) isn't a
particularly well-supported path, for reasons which are generally
sensible:
https://docs.sqlalchemy.org/en/14/faq/sqlexpressions.html#rendering-bound-parameters-inline

However, being able to dump Data Builder queries as syntactically valid,
copy-pastable SQL is such a useful feature that it's worth enduring a
bit of pain here to make this work. (At least, that's the story I'm
telling myself having endured the pain anyway.)

We work around this by compiling the query to a string containing
parameter placeholders and then doing the value substitution ourselves.