Add `regex_utils` module
This allows us to generate random strings matching an arbitrary regular
expression (or at least, one using a reasonable subset of Python's
regular expression dialect).

All the fiddly parsing work is handled for us by Python: we just map the
parsed expression to a small handful of classes which know how to
generate random strings for the element of the expression which they
represent.