fix: Generate random strings
I used to think this was pointless and instead just returned an empty
string. But I now think this has value in making it more immediately
obvious what the problem is: it's not that the dummy data generator
isn't generating anything for the column, it's that it doesn't know
enough about the column to generate anything more helpful.