fix: Display warning for slow running fixtures
Some fixtures involve pulling down Docker images and starting them which
means they can be quite slow the first time they're run. Because of
pytest's output capturing the user has no feedback as to what's
happening and it can look a bit like the tests are broken.

This PR adds a warning message explaining what's happening after a few
seconds delay. Once the relevant containers are started these fixtures
tend to complete very quickly, in which case no message will be
displayed and the output will be exactly as normal.

Closes #723