Ensure `data_strategy` example reprs can round-trip
We want to be able to copy/paste the failing examples Hypothesis
generates for us back into the test module, and so we need to ensure
that their reprs evaluate back to the original value.

This is therefore a meta test, which we don't generally like. But
ensuring that this all works with the dynamically generated classes in
`data_strategy` involves some fairly obnoxious Python which we've
already broken once, so I think it's worth adding a test for this.