Replace `setup-just` action with inline bash
The `setup-just` action regularly, but unpredictably, fails with an API
limit error:

    Error: API rate limit exceeded for 157.55.189.38. (But here's the
    good news: Authenticated requests get a higher rate limit. Check out
    the documentation for more details.)

The documented remedy is to pass the action your GITHUB_TOKEN so it can
use it, but I disagree that it should making these kind of API requests
in the first place.

There's some discussion of the issue here:
https://bennettoxford.slack.com/archives/C63UXGB8E/p1671543684567039

I think longer term we want our own "install stuff" action as outlined
in the thread, but the intermitted failures have got so annoying that I
just want to make them go away for now.