Refactor spec tests to use new test setup machinery
This means that the spec tests are no longer tied to the particular
tables defined in `tests/spec/tables.py` and could be written against
different tables with different structures if that turned out to be
necessary.

As a side effect, because we no longer have to create and teardown
`event_level_table` for tests which only use `patient_level_table`, the
spec tests become noticably faster (27s to 20s, excluding the glacial
Spark tests).