Mark Spark-related code as conditionally covered
This uses the [conditional-coverage][1] pytest plugin to define a new
pragma: `cover-spark-only`. Lines marked with this are not expected to be
covered when pytest is run with a `-k 'not spark'` keyword selector.

It might be slightly surprising that the `setup_queries` loop in
`base_sql.py` is no longer covered. This is because the MSSQL engine has
it's own implementation of this to handle resumable, batched downloads.
And the SQLite engine uses CTEs and therefore has no setup queries.

I expect this path to eventually get excerised by even the SQLite engine
as we add features like uploading patient data from a file.

[1]: https://github.com/wemake-services/coverage-conditional-plugin