feat: Expose `Household`/`HouseholdMember` tables
There's an awkwardness here in that the underlying table is named
generically but in fact contains a snapshot of household membership
computed for us by TPP at a specific point in time. In Cohort Extractor
we dealt with this by accepting a reference date but rejecting all but a
single, [specific value][1].

Here we give the table a name which makes clear that it's tied to a
specific point in time.

I have confirmed in the database that the membership table has only one
row per patient.

[1]: https://github.com/opensafely-core/cohort-extractor/blob/2287eae8/cohortextractor/tpp_backend.py#L2350-L2352