feat: Add `Regex` constraint
This opens up two possibilities:

 * Validating that the real data conforms to the format we have
   documented.

 * Improving the dummy data generation by only generating values which
   conform to the documented format.

In support of both of these we constrain the regular expression language
used to a basic subset of features. This ensures that we know how to
generate matching strings, and should also make it easier to translate
these expressions into whatever dialects we need for our different
database backends when we come to validating the real data.