Refactor `ColumnInfo` to pass constraints through
This means that we don't need to teach the `ColumnInfo` class about each
new type of constraint we add. It also means that we don't need the
`TableInfo` class to hold a reference to the original `TableSchema`
because now that we have the constraints we have everything we need to
build the new schema.