fix: Fix dummy data generation for `date_of_death`
When `date_of_death` is None, as it is allowed to be, we shouldn't attempt
to apply the `FirstOfMonth` constraint to it.

There's a separate question as to whether this column _ought_ to have
this constraint applied: I didn't think these dates were rounded as
dates of birth were. But I'll confirm that separately and create another
PR if necessary.

Closes #969