Notify Slack on generative test failures
I created a new Slack app for this purpose, Data Builder Bot, because I
wanted far more limited privileges than Bennett Bot has. I just want to
be able to write to channels and not read anything. It's possible that
we'll have a more general need for a write-only bot, in which case we
can (hopefully) rename this to something less repo-specific.