fix: New approach to query graph transformation
Instead of force-mutating the state of supposedly immutable objects we
now construct a set of node replacements and then do a single walk over
the query graph to apply them.

The main motivation here was to make it easier to add caching to the
query model's hash methods. But this code also does the right thing in
some cases where the previous code didn't (see later commmits) and I
think it's preferrable to maintain immutability if we can.

This goes some way to addressing #573, though we're not yet using the
functional wonder that is the "zipper".