Test that equal but non-identical Sorts are treated the same
The previous transform code did not pass this test and, as a result, was
generating four times the number of sort queries it needed to when four
values were being selected from a table. It would have been possible to
restructure the ehrQL to avoid this, but I don't think we should have
to.