fix: Faster iteration of deeply nested SQLAlchemy structures
This is the final piece of the current performance puzzle. In #999 we
addressed the performance issues created by a deeply nested stack of
interdependent `GeneratedTable` objects. These are created when running
against a query engine which uses temporary tables, like the MSSQL
engine.

However, an equivalent problem arises when using a query engine that
uses native SQL constructs like CTEs (as the SQLite engine does). Here
the problem shows up when trying to iterate the structure using
SQLAlchemy's supplied functions, which end up recursing back through the
same sections of the graph over and over.

The fix is easy enough: we create a variant of `iterate()` which keeps
track of what it's already seen and avoids re-recursing.