fix: Explicitly CAST date literals to dates
Mostly, MSSQL can infer that they're supposed to be dates and cast them
automatically. But it can't always do this and this has been at the root
of a couple of bugs:
https://github.com/opensafely-core/databuilder/pull/889
https://github.com/opensafely-core/databuilder/issues/998

I think the need for this is down to `pymssql`, which doesn't let us
pass date values directly to the database but insists on formatting them
as strings. So if we switched to a different driver we might not need
this. But I don't think they do any harm, and it has the added benefit
that the debug SQL we generate (which obviously has to only use strings)
behaves correctly as well.