Don't use "magic" SQLAlchemy functions
SQLAlchemy knows about certain common SQL functions and aggregations
(including `MIN`, `MAX`, `SUM` and `COUNT`) and makes these available in
the `sqlalchemy.func` namespace.

However, if you reference a function in this namespace which SQLAlchemy
_doesn't_ know about it will just create one for you with a matching
name but no type information. This is probably quite convenient in some
circumstances, but I don't think we should be making use of these in
Data Builder.

I inadvertently violated my own prohibition against these in a previous
commit, so this PR fixes that.