Add Cloudflare deployment workflow
This is to get a preview of the Data Builder documentation within this
repository, without developers or reviewers having to checkout the
branch themselves.

It has the same limitation as opensafely/documentation#930 — Dependabot
cannot create previews, without extra work.

In practice, this isn't probably much of a problem: the only likely
issues for the documentation are when Material for MkDocs, or MkDocs
have major version changes.

MkDocs has a slow release cycle, and most upgrades for Material for
MkDocs have only required minimal, if any, changes. It is safest for
those to checkout locally and inspect the build.

The other mitigation is adding the `just docs-build` check to the
documentation checks, which will run on Dependabot, because they don't
need additional secrets to run.