Cast only int literals to float and remove support for casting float to int
The main reasons for not supporting cast to float for IntSeries is that we dont think we need it at the moment and it would be easy to add later, but harder to remove. We dont support casting float to int for reasons described in the limitations section of the issue.