Trigger Docker image publication explicitly on tag
The problem this aims to solve is described in #974, particularly this
comment: https://github.com/opensafely-core/databuilder/issues/974#issuecomment-1433087629

Briefly, it seems possible that a tag can be pushed and then not
available in some kind of short time frame, maybe in some kind of race
condition. There is sufficiently short a time between a new tag being
created by the CI workflow and the Docker publication workflow running
to cause the latter workflow to not find the new tag.

(Note: I have looked around for information on what GitHub's behaviour
should be and cannot find what the availability of tags is post-push. So
it is possible there is some other issue we're encountering.)

This approach hopefully works around that problem by setting the trigger
to be the availability of the new tag on `main`.

The downside is that you do lose some flexibility in configuring
triggers based on multiple workflows.

The existing workflow would only trigger publication on a new tag anyway
because it checks for that, so the `workflow_dispatch` trigger here was
redundant.

This also simplifies the workflow, because we don't need to check for
every push whether it has a new matching tag. Instead, we look for tags
and then run the workflow if, and only if, there's a new tag.

We also only tag if all the CI checks pass, so we don't need to ensure
that any other CI-related workflows are completed.