fix: Generate more rows in dummy data tables
This is all utterly unscientific, but without this change I found that
the [study][1] I was working on translating didn't generate enough data for
the analysis scripts to run. Some columns which should have contained
dates ended up completely empty which meant that Stata couldn't guess
the type from the CSV and so the script died with a type error.

Doubling the number of events makes dummy data generation a bit slower,
but solved this specific problem.

[1]: https://github.com/opensafely/bias/pull/3