Simplify test code now that we can rely on `_type`
To be clear, this specific way of checking for `float` would have worked
all along. But while `_type` was only _sometimes_ an accurate guide to
the type a series contains I didn't want to use it and establish
potentially misleading patterns. Now that it's a reliable guide we can
use it freely.