Adapt tests to work with new method
In some ways these tests are a bit redundant now because they're
effectively just retesting features of the in-memory engine which are
tested elsewhere. But we need _some_ level of testing to ensure that our
subclass of the in-memory engine behaves as expected, and I don't want
to have to think about which of these tests to keep. So we'll keep all
of them.