fix: Specify ethnicity variable correctly
`ethnic` doesn't exist as a column, though is included as an item in the
relevant Data Dictionary.

It is actually represented as ten columns, and patients can belong to
multiple groups denoted by these columns.

These are still left as "TODO" values in tests, and generic strings as
represented in Data Builder for now, until we better understand them.