Limit the 'recording' fixture scope to module
The recording fixture is only used in the test_query_model generative
tests. It's used to record all of the generative tests run with
test_query_model and do overall summary reporting and checks.  We don't
want it to be function-scoped (and it can't be with hypothesis anyway),
but if it's session scoped errors don't get reported until the very end
of the test run, and they get reported as an error in the tear-down of
whatever test is last to run, which is confusing.  Setting the scope to
module at least means that the errors is raised within the generative
tests module (and running pytest with -x will stop here, rather than
running until the very end before reporting failed tests)