refactor: Do less manipulation of ISARIC variables
There was some debate over whether these should be nullable Booleans, or
categorical variables (see #1105).

For now, let's have minimal change of what's presented by ISARIC, so
that someone looking at the data dictionary there gets similar results.
We can later review whether we want to do more with this data before a
Data Builder user gets it.

The guidance here:

* Leave existing values as they are, except for `NA` in these
  categorical variables, which corresponds to `NO`
* There should be no other values, and if there are, these will become
  `null`

There will be later checks on the backend in future to ensure the values
as are we expect.