fix: Fix bug in raw SQL compilation
I spotted that `dump-dataset-sql` was blowing up for certain studies and
bisected the issue down to this.

I think this might have been exposed by a change in SQLAlchemy as I'm
pretty sure these studies used to compile to raw SQL without problem.
But in any case, the only reason we have to do all this hand-rolled
parameter interpolation is to work aroud bugs in SQLAlchemy which have
since been fixed in v2.0. Once we get rid of Spark, and thus PyHive, we
can upgrade to SQLAlchemy 2.0 and hopefully throw away a lot of this
code.