Fix literal rendering in our MSSQL Date type
SQLAlchemy has changed such that when a `process_literal_param()` method
is defined this processing is applied as well as, rather than instead
of, the default literal processor. This causes an error because we
return a string, but the default one is expecting a date.

By defining a `literal_processor()` method we can entirely replace the
default literal processing, restoring the old behaviour.