Remove custom SQLAlchemy types
These haven't actually been necessary for a long time as we no longer
rely on SQLAlchemy doing type coercion for us: thanks to the query model
we know exactly what type every value is and whether or not it needs
explicit casting.

The remainder of the `sqlalchemy_types` module should probably be moved
elsewhere but I'm not exactly sure where so I'll leave that task for
another time.