pedantry: Fix relative clauses
Defining relative clauses take "that" and aren't preceded by a comma;
non-defining relative clauses take "which" and are preceded by a comma.
Who says? [English First][1].

"A class that translates between..." is a defining relative clause. "A
class, which translates between..." is a non-defining relative clause.
In this case, my pedantry is less about that/which and more about the
comma.

[1]: https://www.ef.co.uk/english-resources/english-grammar/relative-clauses/