Remove deprecated use of structlog's `wrap_dict`
We aren't calling a logger's `new()`, `bind()` or `unbind()` methods,
so assuming this is unused.

It was originally copied over from cohort-extractor, for which the code
was initially added a while ago.

The current way of doing things if this is needed is to use
`contextvars`:

```python
structlog.contextvars.bind_contextvars()
structlog.contextvars.unbind_contextvars()
```

https://www.structlog.org/en/22.3.0/thread-local.html
https://www.structlog.org/en/22.3.0/contextvars.html