Add new `InsertMany` SQLAlchemy clause
This allows us to say that we'd like a certain collection of rows to be
inserted into a table without specifying the precise mechanics for doing
so. Having this abstraction in place allows us to work independently on
finding the most efficient insert mechanism without needing to change
code elsewhere.