fix: Fix bug in QueryGraphRewriter
After replacing a node we need to pass its replacement through the same
`_rewrite()` method that all other nodes go through. Previously, we
jumped straight to destructuring it and rewriting its attributes. This
is appropriate for some node types but not all. For instance, `Value`
nodes should not be destructured.