Remove `pysqlite-binary` package
The original reason for using this package was that some developers
(particularly those on macOS) had a version of SQLite compiled without
the maths extensions which provide the `FLOOR()` function.

Starting with Python 3.11, the version of SQLite bundled with the macOS
and Windows installers now includes the maths extensions, so this
specific reason no longer applies.

There are some advantages to using an externally provided SQLite whose
version is controlled by us, however:

 * the [`pysqlite-binary`][1] package we were using appears to be
   unmaintained and has no Python 3.11 wheel;
 * the [`pysqlite3`][2] package on which it is based has no plans to
   provide macOS wheels;
 * everything is terrible and we cannot have nice things.

Reverting back to stdlib SQLite resolves the problem for now.

It's possible that if you're developing on an old Linux distribution you
might have a SQLite without maths extenions. But those using old Linux
distributions can generally look after themselves.

[1]: https://github.com/pysqlite3/pysqlite3
[2]: https://github.com/coleifer/pysqlite3