Test against types as they are in the database
Without this change, if we tell SQLAlchemy that a datetime column is a
date and then ask for the contents of this column we will still end up
with `date` objects because SQLAlchemy will use the type information
we've given it to determine the Python type to use for the results. This
is dangerous because the column won't actually _be_ a date, and hence
won't behave like one in comparison operations.

By stripping the type information here we get back the types as they are
in the database which is what we want to be testing against.