fix: Fix region name categories
The previous values didn't match what's actually in the database.
Specifically:

    East of England -> East
    Yorkshire and the Humber -> Yorkshire and The Humber

We also change the order to match the order in which the ONS list these
regions (because we might as well).

The reason these incorrect names crept it was that although Eurostat and
the ONS use the same regional divisions they have _very_ slightly
different labelling. And I must have copied the Eurostat names in early
in the development process without checking against the database.

For reference, here are the ONS names:
https://www.ons.gov.uk/methodology/geography/ukgeographies/eurostat

And here are the Eurostat ones:
https://ec.europa.eu/eurostat/documents/345175/7451602/nuts-map-UK.pdf