Convert empty region/STP to NULL
For the "region" variable this is important because it is typed as a
categorical and the empty string is not one of the defined categories.
This means a file format which is strict about categories (like Arrow)
won't be able to store this value.

As STP is typed as a string this is less critical but we do it for
consistency and also because the empty string does not match the regex
which we define for the column.