Swap `read_dataset_X` functions for `DatasetReader` classes
There are three properties we want when reading dataset files:

 1. validation errors should be raised at the point where the file is
    opened and not the point where it is first iterated;

 2. files should be read in a streaming fashion on demand and not fully loaded
    into memory;

 3. file reader objects should be re-usable in the sense that they can
    be iterated multiple times and will yield the full file contents
    each time.

Points 1 and 2 are in tension to a certain extent – particularly so for
CSV where there's no way of checking that all values are valid without
reading the entire file. By way of compromise we do only basic
validation by checking that the headers are as expected and that all
values in the first 10 rows are of the expected types.

Point 3 is required for handling `InlinePatientTables` in measure
queries, because the measures system requires executing queries multiple
times and so we need to ensure that file reader attached to the
`InlinePatientTable` can be read more than once.

These properties are enforced by the tests added subsequently.