feat: Read then write out user-supplied dummy data
This is in contrast to what we used to do which was attempt to validate
that the user-supplied data exactly matched the format of the real data
and then copy it across unmodified.

By reading in the user-supplied data and writing it out using the same
code path as will be used for real data we can ensure that the output
matches the format that the real data will have while being much more
relaxed about the format of the data we accept.

For instance, we no longer need to care about the order of the columns
in the user-supplied data, or the presence of additional columns. And we
can even allow users to provide dummy data in a different format from
the intended output format (e.g. supply CSV but produce Arrow).