fix: Move `patient_id` check to TableSchema
It's not valid to use `patient_id` as a column on any sort of table, so
it makes sense to validate it at the TableSchema level rather than on a
particular node type.

And given that this is potentially a user-facing, rather than just a
devloper-facing, error we should raise an exception rather than using
`assert`.