Implement maxi/minimum_of
Requires concept of _n_-ary functions and neccesary support in
Hypothesis variable strategies, type checking, query rewriting.

These functions are similar in nature to the ANSI SQL `GREATEST()`
and `LEAST()` functions which are not natively supported in the
currently-used version of mssql or sqlite so engine-specific tricks and
workarounds are required (see https://stackoverflow.com/a/6871572)
which in turn add complexity to the existing logic for aggregate
subqueries and NULL handling.