Reduce example count for long-running generative tests
Until we have time to dig into #1242 properly we're hoping that reducing
the number of examples will avoid triggering the issue while still
getting us value from the generative tests.

If this doesn't work we'll try halving them again (to 10,000) but it's
probably not worth going much further than that.