Use shorter test value to fit column limit
The fact that the column limit is as small as it is suggests that we've
misunderstood somewhere and this column only ever contains a single
code. This doesn't have any particularly terrible consequences
(certainly not compared to the reverse mistake) but it would be good to
get to the bottom of this. See:
https://github.com/opensafely-core/ehrql/issues/1248