Update definition of external test study
This involves pulling in some more ancillary files required by some of
the dataset definitions and updating the list of dataset definitions to
test.

Note that we can't test all of the dataset definitions because some of
them use the `table_from_file` feature and thus require particular
output files to exist before they can be imported.