chore: Only tag versions on `main` branch
We only want to consider tagging the main branch and pull requests.

In principle, this workflow should only be triggered by pushes to main
(because this is configured by `on`), but this is an extra safeguard
should that change.

Recent versions of github-tag-action have slightly confusing behaviour around tagging PRs:

mathieudutour/github-tag-action#150

Confusingly, as that issue mentions, there is an option that specifies
release branches with a *supposed* behaviour:

> **release_branches** (*optional*) - Comma separated list of branches
> (JavaScript regular expression accepted) that will generate the release
> tags. Other branches and pull-requests generate versions postfixed with
> the commit hash and do not generate any repository tag. Examples:
> `master` or `.*` or `release.*,hotfix.*,master`... (default:
> `master,main`).

However, when we tried upgrading to github-tag-action v6.1 before, it was tagging
pull requests, I believe.