Author automated commits as GitHub Actions bot
This uses the name and email as shown in the create-pull-request update
guide for an old version:

https://github.com/peter-evans/create-pull-request/blob/efbb6ba75d718110dc7d3bd08cb98ebd247adc68/docs/updating.md

By default, the create-pull-request action specifies the person who
triggered the workflow as the author.

If this is an automated scheduled run, then the person who is counted as
triggering the workflow is the last person who edited the workflow file.

With #1274, it means that individual authors, may have unsigned and
therefore unverified commits showing up in their commit history.

(I am perhaps fussier about this than other people.)

This is not very obvious behaviour.

By using the GitHub Actions name and email address, we can ascribe the
commits to that, which seems reasonable, as it's entirely automated.