Remove "Contracts" related documentation
To be very clear here (particularly for Seb's benefit!) we are not
abandoning the idea, or even necessarily the terminology, behind
contracts. But I think it's important to maintain conceptual and
terminological consistency in the codebase and documentation; and what
we had previously contained a mixture of asirational and evolving ideas
that didn't directly connect to the reality of how things worked.

I've updated parts of the existing documentation which I think are
shortly going to be replaced anyway (see #1219) because it was easier to
do everything than to worry about what was and wasn't going to be kept.

Depends on #1289