Add wiki-docs content to ehrQL documentation
The ehrQL wiki was used to draft a new and minimal version of the ehrQL
documentation. This was a joint contribution by the Bennett Institute Data
Team, as opposed to solely mine.

(I did contribute, but am merely the committer here.)

This is the content from that wiki.

It will need subsequent work to be properly incorporated into the ehrQL
and OpenSAFELY documentation.