Expose OPA_Diag and OPA_Proc
I made the assumptions that the suffix _Code refers to an ICD10Code for diagnoses (opa_diag) and to an OPCS4Code for procedures (opa_proc) following cohort-extractor. For the suffix _Read I assumed that this refers to a CTV3Code code.

Because there seem to be multiple colums with codes I couldnt follow the usual naming protcol to rename the codes to their specific code system.

We should discuss if we want to expose all of these columns.