Drop the separate data schema page
This was a temporary page explaining where a user could find out about
tables from the ehrQL source code.

We now have a much easier to read reference page. So now, let's refer to
that, but provide a little background context for readers to know that
reference exists, and what it is for.