Test `MedicationDictionary.DMD_ID` is empty string
Although `MedicationDictionary.DMD_ID` is nullable, missing values seem
to be represented as empty strings rather than as `NULL`s. Consequently,
when we `COALESCE` in `TPPBackend.medications()`, we will still prefer
`MedicationDictionary.DMD_ID` to `CustomMedicationDictionary.DMD_ID`,
which is an error. Here, we add a (failing) test to capture the error.