Use better values for test data
Previously, we used values like `"abc"` for test data; fine when we had
a couple of rows of test data, but as the number of rows increased it
become easier and easier to make a copy-paste error. Here, we use values
that approximate the respective coding system but also increment on each
row of test data, which should it make it easier to associate the test's
arrange step with its assert step.