fix: Disable hash seed randomisation
The kinds of DoS attacks hash seed randomisation is designed to protect
against don't apply to ehrQL, and having consistent output makes
debugging much easier. I noticed this when trying to diff the SQL output
from two different versions of study and realised that the order of
codes in a codelist, and the order of columns in some sub-queries,
varied arbitrarily between them.

Unfortunately this can only be done via an environment variable, so the
best we can do is set it automatically for all invocations via Docker or
Just and then warn if ehrQL is invoked via other routes and the variable
is not set.

For more on hash seed randomisation see the note on:
https://docs.python.org/3.11/reference/datamodel.html#object.__hash__