Include Python highlighting in specs code blocks
This has quite a subtle effect when being viewed, but does make the
formatting more consistent throughout the documentation (where I intend
to ensure that all Python code blocks are formatted as such).

It also functions as a semantic indicator for someone viewing the
generated Markdown file as raw text.