Move the errors and examples pages to "how-to"
We are currently following the [Diátaxis approach](https://diataxis.fr/)
to the ehrQL documentation structure.

Initially, the errors and examples pages felt like reference material to
me, when I first structured them like this. However, I'd completely
forgotten that we hadn't got a "how-to" section in the documentation.

On reflection, the errors and examples pages are:

* not tutorials ("lessons that take the reader by the hand through a
  series of steps…" as Diátaxis puts it)
* not really explanatory information ("discussion that clarifies and
  illuminates a particular topic" as Diátaxis puts it), though there may
  be a little context provided, it is primarily concerned with the
  user's immediate task-at-hand
* not an extensive, complete reference (that is, they are not "technical
  descriptions of the machinery" as Diátaxis puts it)
* do give specific guidance to complete specific tasks ("directions that
  take the reader through the steps required to solve a real-world
  problem" as Diátaxis puts it), which are considered "how-to" guides

The fact that these are all listed in one document does not necessarily
make them a reference.

Not that we would want to do this, but in both cases, the sections of
each page could be separated into their own individual pages which are
"how to solve the error…" or "how to extract this information…".

So I think all this justifies these being in a "how-to" section.