feat: Add columns requested in #1358
This adds the columns requested in #1358, with the exception of
`pseudo_visit_id`, which is a binary column. I've checked that columns
of type `bool` have values of `0`, `1`, or `NULL` in the TPP backend. I
haven't mapped from columns of type `int` to their respective
categories.

I've used information from the codebook to map from database column
names to ehrQL column names. To see the mappings side-by-side, see
ehrql/backends/tpp.py.