feat: Add the `ons_cis_raw` table
`ons_cis_raw` contains the columns requested by #1358, with the
exception of the `pseudo_visit_id` column (see below), as well as the
columns of the `ons_cis` table.

I've taken a conservative approach to mapping from the database table
columns (`ONS_CIS_New`) to the ehrQL table (`ons_cis_raw`) columns: I've
lower-cased the former to give the latter.

I'm adding `ons_cis_raw` because we need to do more exploration of the
raw data before modifying `ons_cis` (e.g. checking the values of
`visit_num` are drawn from the interval `[0-15]`), but it's hard to do
more exploration without exposing the raw data through ehrQL.

`pseudo_visit_id` is not included, because `ONS_CIS_New.pseudo_visit_id`
is of type `VARBINARY` and it's not clear how to expose that type
through ehrQL.

Closes #1358