fix: Fix logic for identifying persistent tables
Previously this was entirely backwards. This had the effect that, on
encountering a download error, we would try to close and re-establish
the database connection when using ephemeral tables and _not_
do so when using persistent tables.

This didn't show up in tests because we only simulated errors against
persistent tables, and failing to attempt to re-establish the connection
didn't stop the queries from working (there was nothing actually wrong
with the connection because the errors were just simulated). Now we
simulate errors for both ephemeral and persistent tables which exposes
the bug.

This didn't show up in production because usage hasn't thus far been
heavy enough that we've encountered any download errors.