feat: Expose `ONS_CIS_New.pseudo_visit_id`
Closes #1416

The `ONS_CIS_New.pseudo_visit_id` column is a `varbinary`. Investigation
on L2 tells us the values are 32 bytes (256 bits) in length, so we
assume the values are SHA-256 hashes.

When we expose this column, we want the values to be intelligible
strings. Consequently, when we convert from a `varbinary` to a
`varchar`, we also convert each binary character to two hex characters
by passing `2` as the third positional argument to `CONVERT`. The T-SQL
documentation for [binary styles][1] makes this clear, if you read it
enough:

> 1, 2 A data_type of type character requires a binary expression.
> Each binary character is converted into two hexadecimal characters.
> The characters 0x aren't added to the left of the converted result
> for style 2.

Because of this conversion, we expose this column as `ons_cis.visit_id`
rather than `ons_cis_raw.visit_id`.

[1]: https://learn.microsoft.com/en-us/sql/t-sql/functions/cast-and-convert-transact-sql?view=sql-server-2017#binary-styles