Replace `reformat_docstring` with `strip_indent`
There is a slight behavioural change here in that `reformat_docstring`
would silently swallow `None` and `strip_indent` doesn't. I think this
is a benefit because in most contexts where we use this it's an error
for a docstring not to be defined. And in the few cases where we allow
it we can explicitly handle the `None`.