fix: Set `environ` when generating measures
We did this when generating a dataset, but not when generating measures.
Consequently, environment variables such as `TEMP_DATABASE_NAME` were
available to subclasses of `BaseBackend` when generating datasets but,
not when generating measures.

Users, like @alschaffer, who included the `medications` table in their
measures definitions would have seen references to
`PLACEHOLDER_FOR_TEMP_DATABASE_NAME` in the logs because
`TEMP_DATABASE_NAME` was not passed to `BaseBackend.config`. Now, they
won't.